// FRONTEND/MAINDEX/src/uikit/theme/types.ts
export type Breakpoints = string[]

export type MediaQueries = {
  xs: string
  sm: string
  md: string
  lg: string
  xl: string
  nav: string
}

export type Spacing = number[]

export type Radii = {
  small: string
  default: string
  card: string
  circle: string
}

export type Shadows = {
  level1: string
  active: string
  success: string
  warning: string
  focus: string
  inset: string
  tooltip: string
}

export type Gradients = {
  cardHeader: string
  gold: string
}

export type Colors = {
  primary: string
  secondary: string
  tertiary: string
  success: string
  failure: string
  warning: string
  cardBorder: string
  contrast: string
  dropdown: string
  invertedContrast: string
  input: string
  background: string
  backgroundAlt: string
  backgroundAlt2: string
  text: string
  textDisabled: string
  textSubtle: string
  disabled: string

  // Gradients
  gradients: Gradients

  // Additional colors
  binance: string
  overlay: string
  gold: string
  silver: string
  bronze: string
}

export type ZIndices = {
  dropdown: number
  modal: number
}